Add recycling mechanism for connection IDs
Added a list `closeConnID` in the `ConnManager` struct to store disconnected IDs for recycling. Added the recycling operation in the `NewConnID()` method. When creating a new connection, it would first check if there are any recycled IDs available. If there is at least one, it would use it as new connection's ID. If not, it will increment the current `connID`. This change reduces ID waste and guarantees the uniqueness of connection ID even after many connections are interrupted and reconnected.