Upgrade Go version and refactor `msghandler.go`
Upgraded the Go version in `go.mod` from 1.18 to 1.21 to take advantage of improvements and fixes in the newer version.

In `msghandler.go`, variable name `actual` has been changed to `workQueue` for better readability and understanding of the code. Also, this change brings more clarity while passing channels as arguments in functions.