Add remote configuration and remove redundant instructions from README
This commit introduces the ability to set a remote location for server configuration, enhancing flexibility. The config.go file was adjusted accordingly: a newly added global variable `remoteConfigAddress` stores the address of the remote server; a new function `SetRemoteConfigAddress(address string)` allows to set the value of `remoteConfigAddress`. The initialisation of the server configuration (`InitServerConfig`) was also changed to take into account the potential remote location of the configuration.

Also, some redundant instructions on how to add a YAML file at the root directory were removed from `README.md`. This is a housekeeping change to keep the documentation concise and more manageable.