Refactor connection management and initializations
Connection management was moved from `functional Start()` and `Stop()` to `Start(readerHandler func(), writerHandler func(data []byte))` and `Stop()` in `network/connectionws.go` and `network/connectiontcp.go`. This change ensures the addition and removal of connections to the connection manager happens every time a connection is started and stopped, reducing the chances of connection leaks.

Initialization of various server modules is moved to the `init()` function in `common/base.go` instead of the `main.go` file. This encapsulates the setup code and makes it cleaner to initialize the server, and easier to substitute different implementations if needed.

Unnecessary item space size constants were removed from `common/const.go` as they were not referenced in other parts of the code anymore, and didn't seem to serve any purpose.

jsonsPath and remoteConfigAddress in `config/conf.go` are now local variables, which reduces their scope and makes it clear where they are being used. Global configuration is not fetched remotely until it is requested the first time and remote configuration logic is encapsulated within `initServerConfig()` for better modularity.

`generate_sql.sh` was changed to use the correct filename.

Finally, various general refactorings were performed to make the code cleaner, such as renaming variables and functions to be more descriptive, changing the order of if statements for clarity, removing unused error codes, and combining similar code parts.