Refactoring error handling methods to prevent server crashes. Upgrade proto 2 to proto 3
Modified proto message types from optional to required in .proto files, resulting in a significant number of changes to the compiled .pb.go files. This change was made to simplify the protobuf API and encourage use of simpler primitive types. All field accesses are now safer and more convenient in Go.

Also added a function to capture error messages in api/base.go file. This new function 'MsgErrCapture' captures messages on error and returns a server notification message with an error code and message.

Updated the sql and redis base files to return nil if address is not found in the configuration.

Modified the protocol buffer build script to only support macOS. This simplifies the build process when working on a Mac.

Updated redis and mysql addresses and passwords in the config.toml file removing the hardcoded IP address and passwords to keep code clean and production-ready.