Refactor code and update server config initialization
In this commit, various aspects of the code were refactored for better organization and consistency. The function 'StartOneWorker' was renamed to start with a lower case 's' to reflect Go convention of capitalizing exportable identifiers only. The handling of messages in 'connectionws.go' and 'connectiontcp.go' is now dealt by calling 'GetInstanceMsgHandler()' directly, to facilitate code reusability and to better encapsulate the 'MsgHandler' in MsgHandler module.

The way server configuration is initialized was also updated. The server config is now passed to 'InitServerConfig' as a parameter instead of reading it directly from a file within the function. This provides more flexibility to the code, as the server config can be read from different sources.

An iface file was also renamed from 'inotify.go' to 'ibroadcast.go' for better representation of its purpose. Logging statements were updated to indicate server is starting rather than has started successfully. These changes are aimed at improving code readability, organization, and error handling.